Mlops_projectDocker/requirements.txt
anyio
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asttokens==2.4.0
async-lru==2.0.4
attrs==23.1.0
Babel==2.13.0
backcall==0.2.0
beautifulsoup4==4.6.3
bleach==6.1.0
category-encoders==2.6.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
cliff==2.8.3
cmd2==2.4.3
comm==0.1.4
configparser==6.0.0
contourpy==1.1.1
cssselect==1.0.3
cycler==0.12.1
dacite==1.8.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
exceptiongroup==1.1.3
executing==2.0.0
fastjsonschema==2.18.1
fonttools==4.43.1
htmlmin==0.1.12
idna==3.4
ImageHash==4.3.1
importlib-metadata==6.8.0
importlib-resources==6.1.0
ipykernel==6.25.2
ipython==8.12.3
ipython-genutils==0.2.0
ipywidgets==8.1.1
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-client==8.3.1
jupyter-console==6.6.3
jupyter-core==5.3.2
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter-server==2.7.3
jupyter-server-terminals==0.4.4
jupyterlab==4.0.6
jupyterlab-pygments==0.2.2
jupyterlab-server==2.25.0
jupyterlab-widgets==3.0.9
kaggle==1.5.16
kiwisolver==1.4.5
llvmlite==0.41.0
MarkupSafe==2.1.3
matplotlib==3.7.3
matplotlib-inline==0.1.6
MechanicalSoup==0.8.0
mistune==3.0.2
multimethod==1.10
nbclient==0.8.0
nbconvert==7.9.2
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.1
notebook==7.0.4
notebook-shim==0.2.3
numba==0.58.0
numpy==1.24.4
overrides==7.4.0
packaging==23.2
pandas==2.0.3
pandas-profiling==3.6.6
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.3
pbr==5.11.1
pexpect==4.8.0
phik==0.12.3
pickleshare==0.7.5
Pillow==10.0.1
pkgutil-resolve-name==1.3.10
platformdirs==3.11.0
prettytable==0.7.2
prometheus-client==0.17.1
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydantic==1.10.13
Pygments==2.16.1
pyparsing==3.1.1
pyperclip==1.8.2
python-dateutil==2.8.2
python-json-logger==2.0.7
python-slugify==8.0.1
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.4.4
QtPy==2.4.0
referencing==0.30.2
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.4
scikit-learn==1.3.1
scipy==1.10.1
seaborn==0.12.2
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.3
statsmodels==0.14.0
stevedore==5.1.0
tangled-up-in-unicode==0.2.0
terminado==0.17.1
text-unidecode==1.3
threadpoolctl==3.2.0
tinycss2==1.2.1
tomli==2.0.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.11.2
typeguard==4.1.5
typing-extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
visions==0.7.5
wcwidth==0.2.8
webencodings==0.5.1
websocket-client==1.6.4
widgetsnbextension==4.0.9
wordcloud==1.9.2
ydata-profiling==4.6.0
zipp==3.17.0
fastapi==0.103.2
passlib==1.7.4
uvicorn==0.23.2
pyjwt

